requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
passlib[bcrypt]
PyJWT
python-dotenv
